* intern.h, re.c (rb_memsearch): returns long.
* string.c (rb_str_index): should return offset position.


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@3455 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ intern.h
@@ -330,7 +330,7 @@ VALUE rb_length_by_each _((VALUE));
 /* re.c */
 int rb_memcmp _((char*,char*,long));
 int rb_memcicmp _((char*,char*,long));
-int rb_memsearch _((char*,long,char*,long));
+long rb_memsearch _((char*,long,char*,long));
 VALUE rb_reg_nth_defined _((int, VALUE));
 VALUE rb_reg_nth_match _((int, VALUE));
 VALUE rb_reg_last_match _((VALUE));
